Mostly use f64 for LFO calulations

+++ octasine/src/audio/parameters/common.rs
@@ -13,7 +13,7 @@ pub trait AudioParameter {
     fn set_from_patch(&mut self, value: f64);
     fn get_value_with_lfo_addition(
         &mut self,
-        lfo_addition: Option<f64>,
+        lfo_addition: Option<f32>,
     ) -> <Self::ParameterValue as ParameterValue>::Value;
 
     fn get_parameter_value(&self) -> Self::ParameterValue {
@@ -60,12 +60,12 @@ where
     }
     fn get_value_with_lfo_addition(
         &mut self,
-        lfo_addition: Option<f64>,
+        lfo_addition: Option<f32>,
     ) -> <Self::ParameterValue as ParameterValue>::Value {
         if let Some(lfo_addition) = lfo_addition {
             let patch_value = V::new_from_audio(self.get_value()).to_patch();
 
-            V::new_from_patch((patch_value + lfo_addition).min(1.0).max(0.0)).get()
+            V::new_from_patch((patch_value + lfo_addition as f64).min(1.0).max(0.0)).get()
         } else {
             self.get_value()
         }
@@ -99,10 +99,15 @@ impl<V: ParameterValue> AudioParameter for SimpleAudioParameter<V> {
     }
     fn get_value_with_lfo_addition(
         &mut self,
-        lfo_addition: Option<f64>,
+        lfo_addition: Option<f32>,
     ) -> <Self::ParameterValue as ParameterValue>::Value {
         if let Some(lfo_addition) = lfo_addition {
-            V::new_from_patch((self.patch_value_cache + lfo_addition).min(1.0).max(0.0)).get()
+            V::new_from_patch(
+                (self.patch_value_cache + lfo_addition as f64)
+                    .min(1.0)
+                    .max(0.0),
+            )
+            .get()
         } else {
             self.get_value()
         }
